+++ source/main/utils/setupLogging.js
@@ -36,7 +36,7 @@ export const setupLogging = () => {
 
   log.transports.file.format = (message: Object): string => {
     // Debug level logging is recorded as "info" in Daedalus log files
-    // but in the same time we do not want to output it to console or terminal window
+    // but at the same time we do not want to output it to console or terminal window
     const level = message.level === 'debug' ? 'info' : message.level;
     return formatMessage({ ...message, level });
   };